Fix consumer. Embedded struct collided with method name. Temp code to get consumer working.

+++ internal/ingest/message/message.go
@@ -2,7 +2,7 @@ package message
 
 type Message struct {
 	Data   []byte
-	Offset int
+	Offset int64
 	Source string
 }
